# docs/plans/artifacts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CryptoSizingRequest": {
        "description": "Crypto position sizing calculator.",
        "properties": {
          "entry_price": {
            "exclusiveMinimum": 0.0,
            "title": "Entry Price",
            "type": "number"
          },
          "equity": {
            "exclusiveMinimum": 0.0,
            "title": "Equity",
            "type": "number"
          },
          "leverage": {
            "default": 1.0,
            "minimum": 1.0,
            "title": "Leverage",
            "type": "number"
          },
          "pair": {
            "minLength": 1,
            "title": "Pair",
            "type": "string"
          },
          "risk_pct": {
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Risk Pct",
            "type": "number"
          },
          "side": {
            "enum": [
              "long",
              "short"
            ],
            "title": "Side",
            "type": "string"
          },
          "stop_distance_pct": {
            "exclusiveMinimum": 0.0,
            "title": "Stop Distance Pct",
            "type": "number"
          },
          "taker_fee_bps": {
            "default": 5.0,
            "minimum": 0.0,
            "title": "Taker Fee Bps",
            "type": "number"
          }
        },
        "required": [
          "pair",
          "side",
          "entry_price",
          "equity",
          "risk_pct",
          "stop_distance_pct"
        ],
        "title": "CryptoSizingRequest",
        "type": "object"
      },
...
      "GreeksRequest": {
        "description": "Black-Scholes option Greeks calculation.",
        "properties": {
          "option_type": {
            "enum": [
              "call",
              "put"
            ],
            "title": "Option Type",
            "type": "string"
          },
          "risk_free_rate": {
            "title": "Risk Free Rate",
            "type": "number"
          },
          "spot": {
            "exclusiveMinimum": 0.0,
            "title": "Spot",
            "type": "number"
          },
          "strike": {
            "exclusiveMinimum": 0.0,
            "title": "Strike",
            "type": "number"
          },
          "time_years": {
            "exclusiveMinimum": 0.0,
            "title": "Time Years",
            "type": "number"
          },
          "volatility": {
            "exclusiveMinimum": 0.0,
            "title": "Volatility",
            "type": "number"
          }
        },
        "required": [
          "option_type",
          "spot",
          "strike",
          "time_years",
          "volatility",
          "risk_free_rate"
        ],
        "title": "GreeksRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PipValueRequest": {
        "description": "Compute forex pip value.",
        "properties": {
          "account_conversion_rate": {
            "default": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Account Conversion Rate",
            "type": "number"
          },
          "lot_size": {
            "exclusiveMinimum": 0.0,
            "title": "Lot Size",
            "type": "number"
          },
          "pair": {
            "minLength": 1,
            "title": "Pair",
            "type": "string"
          }
        },
        "required": [
          "pair",
          "lot_size"
        ],
        "title": "PipValueRequest",
        "type": "object"
      },
...
      "ScreenerRunRequest": {
        "description": "Stock screener filter payload.",
        "properties": {
          "chart_points": {
            "default": 252,
            "maximum": 756.0,
            "minimum": 30.0,
            "title": "Chart Points",
            "type": "integer"
          },
          "chart_symbol": {
            "default": "",
            "title": "Chart Symbol",
            "type": "string"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "data_source": {
            "default": "local",
            "title": "Data Source",
            "type": "string"
          },
          "fields": {
            "items": {
              "type": "string"
            },
            "title": "Fields",
            "type": "array"
          },
          "filters": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "number"
              },
              "type": "object"
            },
            "title": "Filters",
            "type": "object"
          },
          "index": {
            "default": "",
            "title": "Index",
            "type": "string"
          },
          "limit": {
            "default": 100,
            "maximum": 2000.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset"
          },
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "percentile_filters": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "number"
              },
              "type": "object"
            },
            "title": "Percentile Filters",
            "type": "object"
          },
          "recommendation": {
            "default": "",
            "title": "Recommendation",
            "type": "string"
          },
          "refresh_cache": {
            "default": false,
            "title": "Refresh Cache",
            "type": "boolean"
          },
          "sector": {
            "default": "",
            "title": "Sector",
            "type": "string"
          },
          "sort_by": {
            "default": "upside_potential",
            "title": "Sort By",
            "type": "string"
          },
          "sort_desc": {
            "default": true,
            "title": "Sort Desc",
            "type": "boolean"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "title": "Symbols",
            "type": "array"
          },
          "technical_scan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technical Scan"
          },
          "template": {
            "default": "",
            "title": "Template",
            "type": "string"
          }
        },
        "title": "ScreenerRunRequest",
        "type": "object"
      },
...
      "StressTestRequest": {
        "description": "Portfolio stress test with factor shocks.",
        "properties": {
          "portfolio_value": {
            "exclusiveMinimum": 0.0,
            "title": "Portfolio Value",
            "type": "number"
          },
          "shocks": {
            "description": "List of {factor, shock_pct, beta} shocks",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "minItems": 1,
            "title": "Shocks",
            "type": "array"
          }
        },
        "required": [
          "portfolio_value",
          "shocks"
        ],
        "title": "StressTestRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "BIST Quant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/professional/crypto-sizing": {
      "post": {
        "description": "Compute crypto position sizing and trade plan.",
        "operationId": "professional_crypto_sizing_api_professional_crypto_sizing_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CryptoSizingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Professional Crypto Sizing Api Professional Crypto Sizing Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Professional Crypto Sizing",
        "tags": [
          "professional"
        ]
      }
    },
    "/api/professional/greeks": {
      "post": {
        "description": "Compute Black-Scholes option Greeks.",
        "operationId": "professional_greeks_api_professional_greeks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GreeksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Professional Greeks Api Professional Greeks Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Professional Greeks",
        "tags": [
          "professional"
        ]
      }
    },
    "/api/professional/pip-value": {
      "post": {
        "description": "Compute forex pip value for a pair and lot size.",
        "operationId": "professional_pip_value_api_professional_pip_value_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipValueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Professional Pip Value Api Professional Pip Value Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Professional Pip Value",
        "tags": [
          "professional"
        ]
      }
    },
    "/api/professional/stress": {
      "post": {
        "description": "Run a portfolio stress test with factor shocks.",
        "operationId": "professional_stress_api_professional_stress_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StressTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Professional Stress Api Professional Stress Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Professional Stress",
        "tags": [
          "professional"
        ]
      }
    },
...
    "/api/screener/run": {
      "post": {
        "description": "Run the stock screener with the given filters.",
        "operationId": "screener_run_api_screener_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenerRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Screener Run Api Screener Run Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Screener Run",
        "tags": [
          "screener"
        ]
      }
    },
    "/api/screener/sparklines": {
      "post": {
        "description": "Return short close-price sparkline series by symbol.\n\nRequest:\n{\n  \"symbols\": [\"THYAO\", \"GARAN\"],\n  \"points\": 60\n}",
        "operationId": "screener_sparklines_api_screener_sparklines_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "items": {
                      "type": "number"
                    },
                    "type": "array"
                  },
                  "title": "Response Screener Sparklines Api Screener Sparklines Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Screener Sparklines",
        "tags": [
          "screener"
        ]
      }
    },
...
  }
}